Move is_async and unwrap_awaitable to utils module, fix method_wrapper obscuring async methods
