+++ apischema/graphql/resolvers.py
@@ -41,9 +41,11 @@ from apischema.typing import get_origin, get_type_hints
 from apischema.utils import (
     get_args2,
     get_origin_or_type2,
+    is_async,
     is_union_of,
     keep_annotations,
     method_registerer,
+    unwrap_awaitable
 )
 from apischema.validation.errors import ValidationError
 
@@ -60,28 +62,6 @@ def partial_serialize(
     obj: Any, *, conversions: HashableConversions = None, aliaser: Aliaser
 ) -> Any:
     return partial_serialization_method(obj.__class__, conversions, aliaser)(obj, False)
-
-
-awaitable_origin = get_origin(Awaitable[Any])
-
-
-def is_async(func: Callable, types: Mapping[str, AnyType] = None) -> bool:
-    if types is None:
-        try:
-            types = get_type_hints(func)
-        except Exception:
-            types = {}
-    return (
-        iscoroutinefunction(func)
-        or get_origin_or_type2(types.get("return")) == awaitable_origin
-    )
-
-
-def unwrap_awaitable(tp: AnyType) -> AnyType:
-    if get_origin_or_type2(tp) == awaitable_origin:
-        return keep_annotations(get_args2(tp)[0] if get_args2(tp) else Any, tp)
-    else:
-        return tp
 
 
 @dataclass(frozen=True)
